Add automated test script for MashTun scoring
