+++ Assets/Scripts/Core/brewing-game-core.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "brewing-game-core"
+}
