--- Assets/Scripts/MashTunViewModel.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MashTunViewModel
-{
-    public string FillButtonLabel;
-    public bool ShowUpgradeButton;    
-    public bool ShowWaterLevelSlider;
-    public float WaterLevelPercentage;   
-    public Color WaterLevelColor;
-    public bool isBrewButtonInteractible;
-    public bool isCollectButtonInteractible;
-    public bool isFillButtonInteractible;
-}
