+++ Assets/Tests/Tests.asmdef
@@ -1,0 +1,12 @@
+{
+    "name": "Tests",
+    "references": [
+        "brewing-game-core"
+    ],
+    "optionalUnityReferences": [
+        "TestAssemblies"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ]
+}
